requirements.txt
discord.py>=2.6.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
asyncio>=3.4.3
jishaku>=2.5.0
pytz>=2023.3
# PostgreSQL dependencies
asyncpg>=0.28.0
SQLAlchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.7
# YAML support for messages
PyYAML>=6.0